smart-vision-classifier/requirements.txt
tensorflow==2.17.0
streamlit==1.39.0
numpy==1.24.3
matplotlib==3.7.2
plotly==5.17.0
Pillow==10.0.1
pandas==2.1.1
scikit-learn==1.3.1
opencv-python==4.8.1.78
seaborn==0.12.2
